improvements
now function define is not a single line, but just a single expression. So now FuncDef also has it's body that was parsed even before builder. Fixed issues with lambdas, they don't have a name so fixed adding them to namespace (they must be just a func on a stack)