improvements & additions
finally, tokenizer+parser are able to parse multi-lined expressions! So now you can split a function body, for example, and make it like in python. The main point is to keep in mind that in the end of the line there MUST be an operator (in case you are out of parenthesis; all the multilined expressions inside of parenthesis must not have an operator in the end)